+++ hooks/post_gen_project.py
@@ -1,52 +1,23 @@
-# post generation step 1: replace relative with absolute path
-# theres no way to get this path while generating the files
-
 from binascii import b2a_base64
 from cookiecutter.utils import work_in
 from hashlib import sha1
 from pathlib import Path
 
-import os
 import random
-import re
 import string
 
 
-target = """{{ cookiecutter.target }}"""
 cwd = Path.cwd()
 basedir = cwd.parent
 
-with work_in(basedir):
-    for dir, subdirs, files in os.walk(cwd / "etc"):
-        for filename in files:
-            infile = Path(dir) / filename
-            lines = []
-            try:
-                with open(infile, "r") as fio:
-                    for line in fio:
-                        mo = re.match(r".*ABSPATH\((.*?)\).*", line)
-                        if mo:
-                            for path in mo.groups():
-                                line = line.replace(
-                                    f"ABSPATH({path})", os.path.abspath(path)
-                                )
-                        lines.append(line)
 
-                with open(infile, "w") as fio:
-                    fio.truncate()
-                    for line in lines:
-                        fio.writelines(line)
-            except Exception:
-                print(f"Can not replace ABSPATH() in file {infile}")
-
-
-# post generation step 2: generate initial user
+# post generation step 1: generate initial user
 username = "{{ cookiecutter.initial_user_name }}" or "admin"
 password = "{{ cookiecutter.initial_user_password }}"
 
-inituser_filename = os.path.join(cwd, "inituser")
+inituser_filename = cwd / "inituser"
 
-if not os.path.exists(inituser_filename):
+if not inituser_filename.exists():
     if not password:
         choices = list(string.ascii_letters + string.digits + "!#$%*(){}[]-")
         random.shuffle(choices)
@@ -57,15 +28,54 @@ if not os.path.exists(inituser_filename):
     with open(inituser_filename, "w") as fp:
         pw = b2a_base64(sha1(password.encode("utf-8")).digest())[:-1]
         fp.write("%s:{SHA}%s\n" % (username, pw.decode("ascii")))
-    os.chmod(inituser_filename, 0o644)
+    inituser_filename.chmod(0o644)
 
 
-# post generation step 3: generate directories
+# post generation step 2: generate directories
 with work_in(basedir):
     Path("{{ cookiecutter.location_clienthome }}").mkdir(parents=True, exist_ok=True)
     Path("{{ cookiecutter.location_log }}").mkdir(parents=True, exist_ok=True)
     Path("{{ cookiecutter.db_blobs_location }}").mkdir(parents=True, exist_ok=True)
-    Path("{{ cookiecutter.environment['CHAMELEON_CACHE'] }}").mkdir(parents=True, exist_ok=True)
+    Path("{{ cookiecutter.environment['CHAMELEON_CACHE'] }}").mkdir(
+        parents=True, exist_ok=True
+    )
     if "{{ cookiecutter.db_storage }}" == "direct":
         filepath = Path("{{ cookiecutter.db_filestorage_location }}")
         filepath.parent.mkdir(parents=True, exist_ok=True)
+    if "{{ cookiecutter.db_storage }}" == "direct":
+        filepath = Path("{{ cookiecutter.db_filestorage_location }}")
+        filepath.parent.mkdir(parents=True, exist_ok=True)
+    if "{{ cookiecutter.db_storage }}" == "relstorage":
+        # import
+        db_relstorage_import_blobs_location = (
+            "{{ cookiecutter.db_relstorage_import_blobs_location | abspath }}"
+        )
+        if db_relstorage_import_blobs_location:
+            filepath = Path(db_relstorage_import_blobs_location)
+            filepath.mkdir(parents=True, exist_ok=True)
+        db_relstorage_import_filestorage_location = (
+            "{{ cookiecutter.db_relstorage_import_filestorage_location | abspath }}"
+        )
+        if db_relstorage_import_filestorage_location:
+            filepath = Path(db_relstorage_import_filestorage_location)
+            filepath.parent.mkdir(parents=True, exist_ok=True)
+
+        # export
+        db_relstorage_export_blobs_location = (
+            "{{ cookiecutter.db_relstorage_export_blobs_location | abspath }}"
+        )
+        if db_relstorage_export_blobs_location:
+            filepath = Path(db_relstorage_export_blobs_location)
+            filepath.mkdir(parents=True, exist_ok=True)
+        db_relstorage_export_filestorage_location = (
+            "{{ cookiecutter.db_relstorage_export_filestorage_location | abspath }}"
+        )
+        if db_relstorage_export_filestorage_location:
+            filepath = Path(db_relstorage_export_filestorage_location)
+            filepath.parent.mkdir(parents=True, exist_ok=True)
+    else:
+        # cleanup relstorage files if no relstorage is configured
+        etc = cwd / "etc"
+        (etc / "relstorage-export.conf").unlink()
+        (etc / "relstorage-import.conf").unlink()
+        (etc / "relstorage-pack.conf").unlink()