+++ CHANGES.rst
@@ -1,14 +1,47 @@
 Changelog
 =========
 
+2.0
+---
+
+- Feature: Implement #14 Generate RelStorage specific pack/import/export configuration.
+  [@jensens, 2024-02-21]
+
+- Feature: Helper script to transform an input configuration file for
+  cookie-cutter-zopeinstance from environment variables with a given prefix
+  and output the result to a file.
+  [@jensens, 2024-01-22]
+
+- Feature: Generate absolute paths in outputs.
+  Fixes https://github.com/plone/cookiecutter-zope-instance/issues/5
+  [@me-kell, @jensens, 2024-01-22]
+
+- Breaking: Flatten zcml variables dict to multiple variables prefixed with zcml.
+  This allows for modification of variables in Docker entry points.
+  [@jensens, 2024-01-21]
+
+- Bugfix: Boolean values in configuration are always true or false (bool).
+  Unified for verbose_security and debug_mode.
+  [@jensens, 2024-01-21]
+
+- Bugfix: typo `db_relstorage_orcale_*` in `cookiecutter.json` fixed.
+  [@mliebischer, @jensens, 2024-02-28]
+
+- Bugfix: `profile_repoze_*` settings were not taken into account.
+  [@mliebischer, @jensens, 2024-02-28]
+
+- Bugfix: Remove unused setting `deprecation_warnings` from `cookiecutter.json`.
+  [@mliebischer, @jensens, 2024-02-28]
+
+
 1.0
 ---
 
-- Add `dos_protection_*` settings for feature introduced in https://github.com/zopefoundation/Zope/pull/1142
-  [jensens, 2023-10-25]
+- Feature: Add `dos_protection_*` settings for feature introduced in https://github.com/zopefoundation/Zope/pull/1142
+  [@jensens, 2023-10-25]
 
-- Fix loading of `overrides.zcml`` and `meta.zcml`.
-  [jensens, 2022-02-07]
+- Bugfix: Loading of `overrides.zcml`` and `meta.zcml`.
+  [@jensens, 2022-02-07]
 
 - Initial work.
-  [jensens]
+  [@jensens]